+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   `Minecraft.mcProfiler`, such as LiteLoader
 - added `AnimatedValue` convenience class
 - added methods to make wireframe `GLUShape`s with arbitrary line widths
+- primitive types are now serializable
 
 ## 1.3
 - updated for Minecraft 1.4.5